Update & bug fixes RT.Java
***Changed***
sysout println changed to print so that when readBackwards(); is invoked it now reads the text backwards and then prints value back on 1 seperate line as a pose to a line for each value. 
 
***Removed***
local ArrayList in pwGen as the method intended to be implemented, has not yet been done.